+++ sushii-2/migrations/20210107212237_create-reminders.sql
@@ -1,0 +1,8 @@
+CREATE TABLE reminders (
+    user_id     BIGINT    NOT NULL,
+    channel_id  BIGINT    NOT NULL,
+    description TEXT      NOT NULL,
+    set_at      TIMESTAMP NOT NULL,
+    expire_at   TIMESTAMP NOT NULL,
+    PRIMARY KEY (user_id, set_at)
+)