+++ sushii-model/src/lib.rs
@@ -3,5 +3,6 @@ pub mod error;
 pub mod keys;
 pub mod model;
 pub mod prelude;
+pub mod utils;
 
 pub use crate::error::{Error, Result};