+++ sushii-model/src/model/sql/mod.rs
@@ -18,5 +18,6 @@ pub use self::{
         cached_user::CachedUser, notification::Notification, user_data::UserData,
         user_level::UserLevel, user_level_global::UserLevelGlobal,
         user_level_ranked::UserLevelRanked, user_xp::UserXP,
+        reminder::Reminder,
     },
 };