+++ sushii-model/src/model/sql/user/mod.rs
@@ -1,5 +1,6 @@
 pub mod cached_user;
 pub mod notification;
+pub mod reminder;
 pub mod user_data;
 pub mod user_level;
 pub mod user_level_global;